+ DB management
#NB: not reponding to ping anymore? /!\

+++ logger.py
@@ -7,13 +7,18 @@ def get_logger(logger_name):
     logger = logging.getLogger(logger_name)
     logger.setLevel(logging.DEBUG)
 
+
 #    filepath = "../logs/{}_{}.log".format(logger_name, datetime.now().strftime('%Y%m%d_%H%M'))
     filepath = "../logs/{}_{}.log".format(logger_name, datetime.now().strftime('%Y%m%d'))
-    os.remove(filepath)
+    if os.path.isfile(filepath):
+        print "Logfile for {} already existing, will try to delete it...".format(logger_name)
+        os.remove(filepath)
+        print "Logfile for {} deleted".format(logger_name)
+
     handler = logging.FileHandler(filepath)
     handler.setLevel(logging.DEBUG)
 
-    formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
+    formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
     handler.setFormatter(formatter)
 
     logger.addHandler(handler)
@@ -22,7 +27,7 @@ def get_logger(logger_name):
     std_out_handler.setFormatter(formatter)
     logger.addHandler(std_out_handler)
 
-    logger.debug('Logger initiated')
+    logger.debug('Logger {} initiated'.format(logger_name))
 
     return logger
